refactor: move mobile, desktop layout Table, List and Row to separate components
