+++ src/pages/Earn/components/EarnTable/index.ts
@@ -1,0 +1,1 @@
+export * from './EarnTable';
